requirements.txt
# Generated from env.yaml for python -m venv
# Python 3.10+

numpy==2.2.5
pandas==2.3.3
matplotlib==3.10.6
plotly==6.5.0
sqlalchemy==2.0.43
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
PyYAML==6.0.3
requests==2.32.5
beautifulsoup4==4.14.2
xgboost==3.1.2
pillow==12.0.0
adjusttext==1.3.0
ai-data-science-team==0.0.0.9017
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
antropy==0.1.9
arch==8.0.0
async-timeout==4.0.3
authlib==1.6.6
backoff==2.2.1
backports-tarfile==1.2.0
bcrypt==5.0.0
beartype==0.22.9
blinker==1.9.0
build==1.3.0
cachetools==6.2.2
chromadb==1.3.5
click==8.2.1
cloudpickle==3.1.2
coloredlogs==15.0.1
cryptography==46.0.3
cyclopts==4.4.4
databricks-sdk==0.73.0
dataclasses-json==0.6.7
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring-parser==0.17.0
docutils==0.22.4
durationpy==0.10
email-validator==2.3.0
et-xmlfile==2.0.0
fakeredis==2.33.0
fastapi==0.124.0
fastmcp==2.14.2
filelock==3.20.0
flask-cors==6.0.1
flask==3.1.2
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
gitpython==3.1.45
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
google-auth==2.41.1
googleapis-common-protos==1.72.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
grpcio==1.76.0
gunicorn==23.0.0
h2o==3.46.0.1
hf-xet==1.2.0
hmmlearn==0.3.3
holidays==0.86
httplib2==0.31.0
httptools==0.7.1
httpx-sse==0.4.3
huey==2.5.5
huggingface-hub==1.1.7
humanfriendly==10.0
id==1.5.0
importlib-metadata==8.7.0
importlib-resources==6.5.2
iniconfig==2.3.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema-path==0.3.4
jupyterlab-widgets==3.0.16
keyring==25.7.0
kubernetes==34.1.0
langchain-chroma==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.6
langchain-experimental==0.4.0
langchain-mcp-adapters==0.2.1
langchain-mcp==0.2.1
langchain-ollama==1.0.0
langchain-openai==1.1.0
langchain-text-splitters==1.0.0
langchain==1.1.0
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.12
langgraph==1.0.4
langsmith==0.4.53
llvmlite==0.45.1
lupa==2.6
mako==1.3.10
markdown-it-py==4.0.0
marshmallow==3.26.1
mcp==1.25.0
mdurl==0.1.2
missingno==0.5.2
mizani==0.14.3
mlflow-skinny==3.7.0
mlflow-tracing==3.7.0
mlflow==3.7.0
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy-extensions==1.1.0
nh3==0.3.2
numba==0.62.1
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==2.8.1
openapi-pydantic==0.5.1
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.4
ormsgpack==1.12.0
pandas-flavor==0.8.1
pathable==0.4.4
pathvalidate==3.3.1
patsy==1.0.2
plotnine==0.15.1
pluggy==1.6.0
polars-runtime-32==1.35.2
polars==1.35.2
posthog==5.4.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.1
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyarrow==21.0.0
pyasn1-modules==0.4.2
pyasn1==0.6.1
pybase64==1.4.2
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydantic==2.12.5
pydeck==0.9.1
pydocket==0.16.3
pyjwt==2.10.1
pypdf==6.4.0
pyperclip==1.11.0
pypika==0.48.9
pyproject-hooks==1.2.0
pytest==9.0.2
python-dotenv==1.2.1
python-multipart==0.0.21
pytimetk==2.4.1
quarto==0.1.0
ray==2.52.1
readme-renderer==44.0
redis==7.1.0
referencing==0.36.2
regex==2025.11.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich-rst==1.3.2
rich==14.2.0
rsa==4.9.1
shellingham==1.5.4
smmap==5.0.2
sortedcontainers==2.4.0
sqlparse==0.5.4
sse-starlette==3.1.2
starlette==0.50.0
statsmodels==0.14.5
streamlit-flow-component==1.6.1
streamlit==1.51.0
supersmoother==0.4
sweetviz==2.3.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.12.0
timebasedcv==0.3.0
tokenizers==0.22.1
toml==0.10.2
tqdm==4.67.1
tsfeatures==0.4.5
twine==6.2.0
typer-slim==0.20.0
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.3.0
uuid-utils==0.12.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
werkzeug==3.1.4
widgetsnbextension==4.0.15
wrapt==1.17.3
xarray==2025.6.1
xxhash==3.6.0
yarl==1.22.0
youtube-transcript-api==1.2.3
zipp==3.23.0
zstandard==0.25.0
